requirements.txt
pydantic
uvicorn
fastapi
httpx
faker
ruff
pytest
psycopg2-binary
sqlalchemy